Optimized code for larger file systems
Removed the old method of traversing the file system using recursion
with a better optimized and safer stack variant.

Also made the permission looser to not fail with even the slightest
permission error (my bad).

Fixed some USAGE.md things, increased the version of the binary,
tweaked the output text of the binary and fixed the config path
for Windows.